Refactor swagger.
