--- src/swagger.js
@@ -1,10 +1,0 @@
-window.onload = function() {
-  // Build a system
-  const editor = SwaggerEditorBundle({
-    dom_id: "#swagger-editor",
-    layout: "StandaloneLayout",
-    presets: [SwaggerEditorStandalonePreset],
-  });
-
-  window.editor = editor;
-};
