+++ src/ts/swagger/wrap-spec-action.ts
@@ -1,0 +1,14 @@
+export const MyWrapActionPlugin = system => {
+  return {
+    statePlugins: {
+      spec: {
+        wrapActions: {
+          updateSpec: (oriAction, system) => str => {
+            console.log("Changed API definition", str);
+            return oriAction(str);
+          },
+        },
+      },
+    },
+  };
+};
